More cleanup and remove host based testing
