+++ azure-pipeline-templates/dist-cache-e2e.yml
@@ -1,13 +1,8 @@
 # Template: distributed cache E2E on a local kind + Tachyon cluster.
 #
 # Brings up a kind (Kubernetes IN Docker) cluster, deploys the Tachyon
-# cache-server helm chart (pulled directly from an OCI-enabled ACR), exposes
-# each cacheserver pod on localhost via kubectl port-forward, then iterates
-# over one or more blobfuse2 mount variants. For each variant it generates a
-# config from the given template, mounts blobfuse2, runs the standard Go E2E
-# suite via e2e-tests.yml, and unmounts before the next variant. This lets a
-# single expensive cluster setup amortise across multiple dist_cache modes
-# (block/file cache, direct-io on/off, ...).
+# cache-server helm chart (pulled directly from an OCI-enabled ACR), deploys
+# blobfuse2 in-cluster, and runs the dedicated dcache_e2e Go suite.
 #
 # See test/scripts/dcache/PLAN.md for why kind (rather than the minikube path
 # vienna-tachyon publishes for) is our substrate.
@@ -16,34 +11,16 @@
 # so failures are debuggable.
 
 parameters:
-  # Identifier for this cluster invocation. Used to name port-forward PID
-  # files, cluster-log artifacts, and per-cluster step display names.
+  # Identifier for this cluster invocation. Used to name cluster-log
+  # artifacts and per-cluster step display names.
   - name: run_idstring
     type: string
-  # List of mount variants to exercise against the standing cluster. Each
-  # entry is an object with:
-  #   id            - short label used in step display names & artifact names
-  #   conf_template - blobfuse2 gen-test-config template file (in testdata/config)
-  #   config_file   - absolute path where the generated config is written
-  - name: variants
-    type: object
   - name: distro_name
     type: string
   - name: account_name
     type: string
   - name: account_key
     type: string
-  - name: account_type
-    type: string
-  - name: adls
-    type: boolean
-    default: false
-  - name: quick_test
-    type: boolean
-    default: true
-  - name: verbose_log
-    type: boolean
-    default: false
 
 steps:
   # -------------------------------------------------------------------------
@@ -80,25 +57,7 @@ steps:
       CACHE_SERVER_CHART_VERSION: $(CACHE_SERVER_CHART_VERSION)
 
   # -------------------------------------------------------------------------
-  # 4. Start kubectl port-forwards so cacheserver pods are reachable from the
-  #    agent host, and stash the resulting server list in a pipeline variable
-  #    so subsequent gen-test-config runs (which read env vars) can substitute
-  #    { DCACHE_SERVERS } in dist_cache config templates.
-  - script: |
-      set -e
-      $(WORK_DIR)/test/scripts/dcache/expose-cacheserver.sh
-      server_list=$(cat "$DCACHE_SERVER_LIST_FILE")
-      echo "Resolved dist_cache server list: $server_list"
-      echo "##vso[task.setvariable variable=DCACHE_SERVERS]$server_list"
-    displayName: 'DistCache [${{ parameters.run_idstring }}]: Expose cacheserver pods'
-    timeoutInMinutes: 5
-    env:
-      DCACHE_SERVER_LIST_FILE: $(Agent.TempDirectory)/dcache_server_list_${{ parameters.run_idstring }}.txt
-      DCACHE_PORTFORWARD_PIDS_FILE: $(Agent.TempDirectory)/dcache_portforward_pids_${{ parameters.run_idstring }}.txt
-
-  # -------------------------------------------------------------------------
-  # 4b. Deploy the in-cluster blobfuse2 pod driven by dcache_e2e (step 5b).
-  #     Host-mount variants in step 5 don't use it.
+  # 4. Deploy the in-cluster blobfuse2 pod driven by dcache_e2e.
   - script: |
       set -e
       $(WORK_DIR)/test/scripts/dcache/deploy-blobfuse2.sh
@@ -112,63 +71,13 @@ steps:
       STO_ACC_CONTAINER: $(containerName)
 
   # -------------------------------------------------------------------------
-  # 5. For each mount variant: gen-config -> mount -> E2E -> unmount.
-  #    The kind cluster is left running across variants; only blobfuse2 is
-  #    remounted between them.
-  - ${{ each variant in parameters.variants }}:
-    - script: |
-        cd $(WORK_DIR)
-        $(WORK_DIR)/blobfuse2 gen-test-config --config-file=${{ variant.conf_template }} --container-name=$(containerName) --temp-path=$(TEMP_DIR) --output-file=${{ variant.config_file }}
-      displayName: 'DistCache [${{ variant.id }}]: Create config file'
-      env:
-        STO_ACC_NAME: ${{ parameters.account_name }}
-        STO_ACC_KEY: ${{ parameters.account_key }}
-        STO_ACC_TYPE: ${{ parameters.account_type }}
-        STO_ACC_ENDPOINT: 'https://${{ parameters.account_name }}.blob.core.windows.net'
-        USE_HTTP: 'false'
-        VERBOSE_LOG: ${{ parameters.verbose_log }}
-        DCACHE_SERVERS: $(DCACHE_SERVERS)
-      continueOnError: false
-
-    - script: |
-        cat ${{ variant.config_file }}
-      displayName: 'DistCache [${{ variant.id }}]: Print config file'
-
-    - template: './e2e-tests.yml'
-      parameters:
-        adls: ${{ parameters.adls }}
-        idstring: 'DistCache_${{ variant.id }}'
-        distro_name: ${{ parameters.distro_name }}
-        quick_test: ${{ parameters.quick_test }}
-        artifact_name: '${{ parameters.distro_name }}_DistCache_${{ variant.id }}.txt'
-        verbose_log: ${{ parameters.verbose_log }}
-        mountStep:
-          script: |
-            $(WORK_DIR)/blobfuse2 mount $(MOUNT_DIR) --config-file=${{ variant.config_file }} --default-working-dir=$(System.DefaultWorkingDirectory)
-          displayName: 'DistCache [${{ variant.id }}]: Mount'
-          timeoutInMinutes: 3
-          continueOnError: false
-
-    # Unmount blobfuse2 (but keep the Azure container + kind cluster) so the
-    # next variant starts from a clean mount point.
-    - template: 'cleanup.yml'
-      parameters:
-        unmount: true
-        delete_containers: false
-
-  # -------------------------------------------------------------------------
-  # 5b. Run the pod-focused suite once after all host-mount variants. Its pod
-  # uses the fixed discovery-based block-cache config deployed in step 4b, so
-  # rerunning it per host variant would not add variant coverage. Running it
-  # last also ensures its kind-node fault test cannot break a port-forward
-  # needed by a later host mount.
+  # 5. Run only the distributed-cache-specific pod suite.
   - script: |
       set -e
       cd $(WORK_DIR)
       go test -v -timeout=30m -tags=fuse3 ./test/dcache_e2e/... -args \
         -pod-namespace=blobfuse2-dist-cache \
         -pod-deployment=blobfuse2-dist-cache \
-        -pod-selector=app=blobfuse2-dist-cache \
         -pod-mount-path=/mnt/blobfuse_mnt \
         -kubectl-bin=kubectl \
         -docker-bin=docker \
@@ -208,5 +117,3 @@ steps:
     displayName: 'DistCache [${{ parameters.run_idstring }}]: Teardown kind cluster'
     condition: always()
     timeoutInMinutes: 15
-    env:
-      DCACHE_PORTFORWARD_PIDS_FILE: $(Agent.TempDirectory)/dcache_portforward_pids_${{ parameters.run_idstring }}.txt
