+++ test/dcache_e2e/README.md
@@ -1,20 +1,17 @@
 # dist_cache E2E tests (`test/dcache_e2e/`)
 
-Read-path E2E tests focused on the `dist_cache` component. Complements
-`test/e2e_tests/` (which is generic filesystem correctness under a
-`dist_cache`-enabled config): the tests here specifically assert
+Read-path E2E tests focused on the `dist_cache` component. These tests assert
 `dist_cache` *behaviour* — L2 miss populates, L2 hit serves the same bytes
-we wrote, cache-server metrics move in the expected direction.
+we wrote, and cache-server metrics move in the expected direction.
 
 ## Design
 
-* **Mount is read-only.** `dist_cache` is a read-only L2 in the current
-  scope; writes never traverse it. The tests mount blobfuse2 with
-  `--read-only=true` and seed all payloads directly to Azure Storage via
-  the Azure SDK.
-* **Test owns the mount lifecycle.** To distinguish an L2-hit read from a
-  same-handle in-memory read, the test unmounts and remounts between reads
-  so that `block_cache`'s cooked blocks are guaranteed gone.
+* **The mount is read-only.** The reference pod config sets `read-only: true`,
+  and every test verifies the cloned pod reports `ro` in `/proc/mounts` before
+  reading. Payloads are uploaded directly to Azure Storage through the Azure SDK.
+* **Each test owns an isolated Deployment.** A test clones the reference
+  blobfuse2 Deployment with a unique name and selector, then deletes it in
+  cleanup. Tests that need to clear local `block_cache` restart their clone.
 * **Behaviour + metrics.** Data integrity is asserted by MD5 on every
   read. Cache behaviour is asserted by diffing Prometheus counters from
   the Tachyon cache-server pods between snapshots. When metric endpoints
@@ -26,38 +23,36 @@ we wrote, cache-server metrics move in the expected direction.
 | File | Purpose |
 |---|---|
 | `main_test.go` | `TestMain`, flag registration, per-run config. |
-| `mount.go` | Mount / unmount / remount helpers (owns the FUSE lifecycle). |
-| `helpers.go` | Random payload + MD5 + Azure SDK upload / delete / download. |
+| `pod_mount_test.go` | Per-test Deployment cloning, scaling, restart, and mount-read helpers. |
+| `helpers_test.go` | Random payload, MD5, and Azure SDK upload/delete helpers. |
 | `metrics_test.go` | In-pod Prometheus scraper (`kubectl exec ... curl`) and `CacheServerMetrics` delta helpers. |
+| `metrics_parse_test.go` | Unit tests for Prometheus parsing and metric deltas. |
 | `read_path_test.go` | `TestReadPath_L2MissPopulatesAndHits` — the canonical L2 miss → populate → hit test. |
 | `node_failure_test.go` | Stops a kind worker hosting one cache-server, verifies mixed L2/Azure fallback, then restores the node. |
+| `stampede_test.go` | Verifies concurrent cold reads coalesce into one Azure GET. |
+| `warm_cache_test.go` | Verifies late-joining pods read previously populated data from L2. |
 
 ## Running locally
 
 Prerequisites (in addition to the standard `blobfuse2` build):
 
-1. A kind cluster with Tachyon deployed and its wire port reachable on
-   `localhost` via `kubectl port-forward`. Use
+1. A kind cluster with Tachyon deployed. Use
    [test/scripts/dcache/setup-kind.sh](../scripts/dcache/setup-kind.sh) +
-   [deploy-tachyon.sh](../scripts/dcache/deploy-tachyon.sh) +
-   [expose-cacheserver.sh](../scripts/dcache/expose-cacheserver.sh).
-   Prometheus scrapes are done in-cluster via `kubectl exec ... curl`,
-   so no separate metrics port-forward is required.
+  [deploy-tachyon.sh](../scripts/dcache/deploy-tachyon.sh).
+  The pod suite uses in-cluster discovery and scrapes Prometheus through
+  `kubectl exec`.
 2. A running blobfuse2 Deployment in the kind cluster. Deploy it with
    [test/scripts/dcache/deploy-blobfuse2.sh](../scripts/dcache/deploy-blobfuse2.sh),
    which renders
    [docker/k8s/blobfuse2-dist-cache-deployment.yaml.tmpl](../../docker/k8s/blobfuse2-dist-cache-deployment.yaml.tmpl)
    with your storage credentials + image ref, side-loads the image into
    every kind node, and waits for rollout. The defaults expect deployment
-   `blobfuse2-dist-cache` in namespace `blobfuse2-dist-cache`, selected
-   by `app=blobfuse2-dist-cache`, with its FUSE mount at
+  `blobfuse2-dist-cache` in namespace `blobfuse2-dist-cache`, with its FUSE mount at
    `/mnt/blobfuse_mnt` inside the container.
 3. An Azure Storage account + container the test can seed data into.
 
 ```bash
 # From the repo root, with blobfuse2 already built.
-
-export DCACHE_SERVERS=$(cat /tmp/dcache_server_list.txt)
 
 # Seed the credentials the test uses to upload payloads out-of-band AND
 # that deploy-blobfuse2.sh bakes into the pod's config Secret.
@@ -77,7 +72,6 @@ go test -v -tags=fuse3 ./test/dcache_e2e/... \
   -run '^TestReadPath_L2MissPopulatesAndHits$' -args \
   -pod-namespace=blobfuse2-dist-cache \
   -pod-deployment=blobfuse2-dist-cache \
-  -pod-selector=app=blobfuse2-dist-cache \
   -pod-mount-path=/mnt/blobfuse_mnt \
   -kubectl-bin=kubectl
 ```
@@ -99,6 +93,4 @@ Tachyon Helm chart exposes it somewhere other than 9096.
 
 Wired from
 [azure-pipeline-templates/dist-cache-e2e.yml](../../azure-pipeline-templates/dist-cache-e2e.yml).
-For each mount variant the template already exercises via
-`e2e-tests.yml`, an additional `dcache_e2e Go tests` step runs the tests
-in this package against the same generated config.
+The template deploys the in-cluster dependencies and runs only this package.
