+++ test/scripts/dcache/config/nightly.config
@@ -18,7 +18,6 @@ PREREQ_RELEASE_NAME="${PREREQ_RELEASE_NAME:-cache-server-prereq}"
 KIND_NODES="${KIND_NODES:-4}"
 KIND_VERSION="${KIND_VERSION:-v0.32.0}"
 KIND_NODE_IMAGE="${KIND_NODE_IMAGE:-kindest/node:v1.36.1}"
-KIND_CLUSTER_CONFIG="${KIND_CLUSTER_CONFIG:-}"   # populated by setup-kind.sh
 
 CACHE_SERVER_PORT="${CACHE_SERVER_PORT:-9065}"
 CACHE_SERVER_REPLICAS="${CACHE_SERVER_REPLICAS:-3}"
