+++ test/scripts/dcache/teardown-kind.sh
@@ -16,23 +16,6 @@ if [[ -f "$CONFIG_FILE" ]]; then
 fi
 
 CLUSTER_NAME="${CLUSTER_NAME:-blobfuse-dcache}"
-DEFAULT_OUTPUT_DIR="${AGENT_TEMPDIRECTORY:-/tmp}"
-DCACHE_PORTFORWARD_PIDS_FILE="${DCACHE_PORTFORWARD_PIDS_FILE:-$DEFAULT_OUTPUT_DIR/dcache_portforward_pids.txt}"
-
-# --- Kill port-forwards ----------------------------------------------------
-kill_pids_from_file() {
-    local pids_file="$1"
-    [[ -f "$pids_file" ]] || return 0
-    echo "Killing port-forward PIDs from $pids_file ..."
-    while read -r pid; do
-        [[ -z "$pid" ]] && continue
-        echo "  kill $pid"
-        kill "$pid" 2>/dev/null
-    done < "$pids_file"
-    rm -f "$pids_file"
-}
-
-kill_pids_from_file "$DCACHE_PORTFORWARD_PIDS_FILE"
 
 # --- Delete cluster --------------------------------------------------------
 # `kind delete cluster` destroys the node containers wholesale, so per-release
